+++ src/Chunkyard/Options.cs
@@ -1,17 +1,41 @@
 namespace Chunkyard;
 
+public abstract class Options
+{
+    protected Options(
+        string repository,
+        int parallel,
+        Prompt prompt)
+    {
+        Repository = repository;
+        Parallel = parallel;
+        Prompt = prompt;
+    }
+
+    [Option('r', "repository", Required = true, HelpText = "The repository path.")]
+    public string Repository { get; }
+
+    [Option("parallel", Required = false, HelpText = "The degree of parallelism.", Default = -1)]
+    public int Parallel { get; }
+
+    [Option("prompt", Required = false, HelpText = "The password prompt method.", Default = Prompt.Console)]
+    public Prompt Prompt { get; }
+}
+
 [Verb("cat", HelpText = "Export or print the value of a snapshot or a set of chunk IDs.")]
-public sealed class CatOptions
+public sealed class CatOptions : Options
 {
     public CatOptions(
         IEnumerable<string> chunkIds,
         int snapshotId,
+        string? export,
         string repository,
-        string? export)
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
         ChunkIds = chunkIds;
         SnapshotId = snapshotId;
-        Repository = repository;
         Export = export;
     }
 
@@ -21,30 +45,26 @@ public sealed class CatOptions
     [Option('s', "snapshot", Required = false, HelpText = "The snapshot ID.", Default = SnapshotStore.LatestSnapshotId, SetName = "snapshots")]
     public int SnapshotId { get; }
 
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
-
     [Option('e', "export", Required = false, HelpText = "The export path.", Default = "")]
     public string? Export { get; }
 }
 
 [Verb("check", HelpText = "Check if a snapshot is valid.")]
-public sealed class CheckOptions
+public sealed class CheckOptions : Options
 {
     public CheckOptions(
-        string repository,
         int snapshotId,
         IEnumerable<string> includePatterns,
-        bool shallow)
+        bool shallow,
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
-        Repository = repository;
         SnapshotId = snapshotId;
         IncludePatterns = includePatterns;
         Shallow = shallow;
     }
-
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
 
     [Option('s', "snapshot", Required = false, HelpText = "The snapshot ID.", Default = SnapshotStore.LatestSnapshotId)]
     public int SnapshotId { get; }
@@ -57,43 +77,41 @@ public sealed class CheckOptions
 }
 
 [Verb("copy", HelpText = "Copy snapshots from one repository to another.")]
-public sealed class CopyOptions
+public sealed class CopyOptions : Options
 {
     public CopyOptions(
-        string sourceRepository,
-        string destinationRepository)
+        string destinationRepository,
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
-        SourceRepository = sourceRepository;
         DestinationRepository = destinationRepository;
     }
-
-    [Option('s', "source", Required = true, HelpText = "The source repository path.")]
-    public string SourceRepository { get; }
 
     [Option('d', "destination", Required = true, HelpText = "The destination repository path.")]
     public string DestinationRepository { get; }
 }
 
 [Verb("store", HelpText = "Store a new snapshot.")]
-public sealed class StoreOptions
+public sealed class StoreOptions : Options
 {
     public StoreOptions(
         IEnumerable<string> paths,
-        string repository,
         IEnumerable<string> includePatterns,
-        bool preview)
+        bool preview,
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
         Paths = paths;
-        Repository = repository;
         IncludePatterns = includePatterns;
         Preview = preview;
     }
 
     [Option('p', "paths", Required = true, HelpText = "The files and directories (blobs) to store.")]
     public IEnumerable<string> Paths { get; }
-
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
 
     [Option('i', "include", Required = false, HelpText = "The fuzzy patterns for blobs to include.")]
     public IEnumerable<string> IncludePatterns { get; }
@@ -103,24 +121,23 @@ public sealed class StoreOptions
 }
 
 [Verb("diff", HelpText = "Show the difference between two snapshots.")]
-public sealed class DiffOptions
+public sealed class DiffOptions : Options
 {
     public DiffOptions(
-        string repository,
         int firstSnapshotId,
         int secondSnapshotId,
         IEnumerable<string> includePatterns,
-        bool chunksOnly)
+        bool chunksOnly,
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
-        Repository = repository;
         FirstSnapshotId = firstSnapshotId;
         SecondSnapshotId = secondSnapshotId;
         IncludePatterns = includePatterns;
         ChunksOnly = chunksOnly;
     }
-
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
 
     [Option('f', "first", Required = false, HelpText = "The first snapshot ID.", Default = SnapshotStore.SecondLatestSnapshotId)]
     public int FirstSnapshotId { get; }
@@ -136,84 +153,81 @@ public sealed class DiffOptions
 }
 
 [Verb("gc", HelpText = "Remove unreferenced chunks.")]
-public sealed class GarbageCollectOptions
+public sealed class GarbageCollectOptions : Options
 {
-    public GarbageCollectOptions(string repository)
+    public GarbageCollectOptions(
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
-        Repository = repository;
     }
-
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
 }
 
 [Verb("keep", HelpText = "Keep only the given list of snapshots.")]
-public sealed class KeepOptions
+public sealed class KeepOptions : Options
 {
     public KeepOptions(
         int latestCount,
-        string repository)
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
         LatestCount = latestCount;
-        Repository = repository;
     }
 
     [Option("latest", Required = true, HelpText = "The count of the latest snapshots to keep.")]
     public int LatestCount { get; }
-
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
 }
 
 [Verb("list", HelpText = "List all snapshots.")]
-public sealed class ListOptions
+public sealed class ListOptions : Options
 {
-    public ListOptions(string repository)
+    public ListOptions(
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
-        Repository = repository;
     }
-
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
 }
 
 [Verb("remove", HelpText = "Remove a snapshot.")]
-public sealed class RemoveOptions
+public sealed class RemoveOptions : Options
 {
     public RemoveOptions(
         int snapshotId,
-        string repository)
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
         SnapshotId = snapshotId;
-        Repository = repository;
     }
 
     [Option('s', "snapshot", Required = true, HelpText = "The snapshot ID.")]
     public int SnapshotId { get; }
-
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
 }
 
 [Verb("restore", HelpText = "Restore a snapshot.")]
-public sealed class RestoreOptions
+public sealed class RestoreOptions : Options
 {
     public RestoreOptions(
-        string repository,
         string directory,
         int snapshotId,
         IEnumerable<string> includePatterns,
-        bool preview)
+        bool preview,
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
-        Repository = repository;
         Directory = directory;
         SnapshotId = snapshotId;
         IncludePatterns = includePatterns;
         Preview = preview;
     }
-
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
 
     [Option('d', "directory", Required = false, HelpText = "The directory to restore into.", Default = ".")]
     public string Directory { get; }
@@ -229,22 +243,21 @@ public sealed class RestoreOptions
 }
 
 [Verb("show", HelpText = "Show the content of a snapshot.")]
-public sealed class ShowOptions
+public sealed class ShowOptions : Options
 {
     public ShowOptions(
-        string repository,
         int snapshotId,
         IEnumerable<string> includePatterns,
-        bool chunksOnly)
+        bool chunksOnly,
+        string repository,
+        int parallel,
+        Prompt prompt)
+        : base(repository, parallel, prompt)
     {
-        Repository = repository;
         SnapshotId = snapshotId;
         IncludePatterns = includePatterns;
         ChunksOnly = chunksOnly;
     }
-
-    [Option('r', "repository", Required = false, HelpText = "The repository path.", Default = Commands.DefaultRepository)]
-    public string Repository { get; }
 
     [Option('s', "snapshot", Required = false, HelpText = "The snapshot ID.", Default = SnapshotStore.LatestSnapshotId)]
     public int SnapshotId { get; }